Further moving the driver classes from the old directory to the new one.
