+++ storage/database/drivers/mysqlpdo/Restriction.php
@@ -1,0 +1,29 @@
+<?php namespace spitfire\storage\database\drivers\mysqlpdo;
+
+use spitfire\storage\database\Restriction as AbstractRestriction;
+
+class Restriction extends AbstractRestriction
+{
+	public function __toString() {
+		$value = $this->getValue();
+
+		if (is_array($value)) {
+			foreach ($value as &$v) {
+				$v = $this->getTable()->getDb()->quote($v);
+			}
+
+			$quoted = implode(',', $value);
+			return "{$this->getField()} {$this->getOperator()} ({$quoted})";
+		}
+
+		elseif ($value instanceof QueryField) {
+			return "{$this->getField()} {$this->getOperator()} {$this->getValue()}";
+		}
+
+		else {
+			$quoted = $this->getTable()->getDb()->quote($this->getValue());
+			return "{$this->getField()} {$this->getOperator()} {$quoted}";
+		}
+	}
+	
+}
